// src/interface/agendamentos/update-agendamento-interface.ts
import { StatusAgendamento } from "./status-agendamento-interface";

export interface IUpdateAppointment {
  id: string;              
  nome?: string;
  telefone?: string;
  email?: string;
  data?: string;
  hora?: string;           
  servico?: string;
  profissional?: string;
  status?: StatusAgendamento;
  inicio?: string;
  fim?: string;
}